+++ src/Maintenance/Maintenance/ChangeLog.md
@@ -19,7 +19,8 @@
 -->
 
 ## Upcoming Release
-
+* Fixed breaking change information
+  
 ## Version 1.3.0
 * Added support for Resource Group and Subscription configuration assignment.